requirements.txt
asgiref==3.3.1
Django==3.1.7
pip==21.0.1
pytz==2021.1
setuptools==54.1.1
sqlparse==0.4.1
django-autoslug==1.9.8

